Split crt.sh handling off into its own module
